backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
celery==5.3.1
redis==4.5.5
sqlalchemy==2.0.19
mysql-connector-python==8.1.0
elasticsearch==8.9.0
pydantic==1.10.12
python-dotenv==1.0.0
passlib
pymysql
kafka-python
scikit-learn
pandas
numpy